final/requirements.txt
functions-framework==3.*
pandas==1.5.3
requests==2.28.2
google-cloud-secret-manager==2.16.2
pandas-gbq==0.19.2
flask